+++ packages/libs/sdk/src/api/types/contracts/getContractDetails.ts
@@ -8,7 +8,7 @@ import {
   isEvmAddress,
   supportedChainInput,
 } from '../../../lib/validation/validators';
-import { SimplifyType } from '../../utils/helpers';
+import { SimplifyType } from '../../../lib/types';
 import { z } from 'zod';
 
 export type ContractDetailsQuery = {
