+++ packages/libs/sdk/src/api/types/events/getAll.ts
@@ -9,7 +9,7 @@ import {
   baseEventsInput,
   supportedChainInput,
 } from '../../../lib/validation/validators';
-import { SimplifyType } from '../../utils/helpers';
+import { SimplifyType } from '../../../lib/types';
 
 import { z } from 'zod';
