+++ packages/libs/sdk/src/api/types/events/getByContract.ts
@@ -10,7 +10,7 @@ import {
   CodegenPaginationFragment,
 } from '../../graphql/generatedTypes';
 import { ChainName } from '../chains';
-import { SimplifyType } from '../../utils/helpers';
+import { SimplifyType } from '../../../lib/types';
 import { z } from 'zod';
 
 export type ContractEventsQuery = {
