+++ packages/libs/sdk/src/api/types/nfts/getByWalletAddress.ts
@@ -12,7 +12,7 @@ import {
   CodegenPaginationFragment,
 } from '../../graphql/generatedTypes';
 import { ChainName } from '../chains';
-import { SimplifyType } from '../../utils/helpers';
+import { SimplifyType } from '../../../lib/types';
 import { z } from 'zod';
 
 // Using the generated CodegenEthMainnetWalletNFTsByAddressQuery as a base for the type here
