+++ packages/libs/sdk/src/api/types/nfts/getCollectionDetails.ts
@@ -8,7 +8,7 @@ import {
   CodegenNftCollectionInfoFragment,
 } from '../../graphql/generatedTypes';
 import { ChainName } from '../chains';
-import { SimplifyType } from '../../utils/helpers';
+import { SimplifyType } from '../../../lib/types';
 import { z } from 'zod';
 
 export type NftCollectionDetailsQuery = {
