+++ packages/libs/sdk/src/api/types/tokens/getBalancesByWalletAddress.ts
@@ -11,7 +11,7 @@ import {
   CodegenPaginationFragment,
 } from '../../graphql/generatedTypes';
 import { ChainName } from '../chains';
-import { SimplifyType } from '../../utils/helpers';
+import { SimplifyType } from '../../../lib/types';
 import { z } from 'zod';
 
 // Using the generated CodegenEthMainnetBalancesByWalletAddressQuery as a base for the type here
