+++ packages/libs/sdk/src/api/types/transactions/getBySearch.ts
@@ -9,7 +9,7 @@ import {
   CodegenPaginationFragment,
 } from '../../graphql/generatedTypes';
 import { ChainName } from '../chains';
-import { SimplifyType } from '../../utils/helpers';
+import { SimplifyType } from '../../../lib/types';
 import { z } from 'zod';
 
 // Using the generated CodegenEthMainnetTransactionsBySearchQuery as a base for the type here
