+++ packages/libs/sdk/src/api/types/transactions/getByWalletAddress.ts
@@ -11,7 +11,7 @@ import {
   CodegenPaginationFragment,
 } from '../../graphql/generatedTypes';
 import { ChainName } from '../chains';
-import { SimplifyType } from '../../utils/helpers';
+import { SimplifyType } from '../../../lib/types';
 import { z } from 'zod';
 
 // Using the generated CodegenEthMainnetTransactionsByWalletAddressQuery as a base for the type here
