+++ packages/libs/sdk/src/api/types/utils/gasPrices.ts
@@ -8,7 +8,7 @@ import {
   CodegenGasPrice,
 } from '../../graphql/generatedTypes';
 import { ChainName } from '../chains';
-import { SimplifyType } from '../../utils/helpers';
+import { SimplifyType } from '../../../lib/types';
 import { z } from 'zod';
 
 export type GasPricesQuery = {
