+++ packages/libs/sdk/src/lib/types/ValueOf.ts
@@ -1,0 +1,8 @@
+/**
+ * @description Creates a type that extracts the values of T.
+ *
+ * @example
+ * ValueOf<{ a: string, b: number }>
+ * => string | number
+ */
+export type ValueOf<T> = T[keyof T];
