+++ packages/libs/sdk/src/lib/types/index.ts
@@ -1,1 +1,2 @@
 export * from './SimplifyType';
+export * from './ValueOf';
